Revert "Workaround pypy/pypy#5252"
This reverts commit dc274c9f4a7b5ef9a573b175d5c06dce3910d7e7.

+++ src/gmpy2_context.c
@@ -185,7 +185,7 @@ static PyObject *
 GMPy_CTXT_Exit(PyObject *self, PyObject *args)
 {
     CTXT_Object *ctx = (CTXT_Object*)self;
-    int res = (int)PyContextVar_Reset(current_context_var, ctx->token);  // XXX: pypy/pypy#5252
+    int res = PyContextVar_Reset(current_context_var, ctx->token);
     Py_DECREF(ctx->token);
     if (res == -1) {
         SYSTEM_ERROR("Unexpected failure in restoring context.");
